Moved listener to a separate class
